album and image manipulation crud refactoring with auth - user login and token generation dashboard

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Models\Album;
 use App\Http\Resources\V1\AlbumResource;
+use Illuminate\Http\Request;
 
 class AlbumController extends Controller
 {
@@ -15,9 +16,9 @@ class AlbumController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate());
+        return AlbumResource::collection(Album::where('user_id', $request->user()->id)->paginate());
     }
 
     /**
@@ -39,8 +40,11 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
+        if ($album->user_id != $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         return new AlbumResource($album);
     }
 
@@ -52,7 +56,10 @@ class AlbumController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateAlbumRequest $request, Album $album)
-    {
+    { 
+        if ($album->user_id != $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         $album->update($request->all());
         
         return new AlbumResource($album);
@@ -64,10 +71,12 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request, Album $album)
     {
+        if ($album->user_id != $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         $album->delete();
-        
         return response('', 204);
     }
 }
